requirements.txt
watchdog==3.0.0
Pillow==10.1.0
pycups==2.0.1
python-dotenv==1.0.0